requirements.txt
Django==1.6.0
selenium==2.39.0
coverage==3.7.1
django-filter==0.7
djangorestframework==2.3.12
Markdown==2.3.1